Renaming App Loger to -> AppLogger
